Light restructure of Ptolemaic.
