+++ everest/ptolemaic/_ptolemaic.py
@@ -14,7 +14,7 @@ class PtolemaicMeta(_ABCMeta):
 
 
 @_classtools.Diskable
-class Ptolemaic(metaclass = PtolemaicMeta):
+class Ptolemaic(metaclass = PtolemaicMeta): # pylint: disable=R0903
 
     @classmethod
     @_abstractmethod
