+++ everest/ptolemaic/datalike/primary/__init__.py
@@ -3,13 +3,14 @@
 ###############################################################################
 
 
-from .. import _classtools
+from .. import _classtools, _everestutilities
 
-from . import _Datalike
+from .. import Datalike as _Datalike
+from ..secondary import functional as _functional
 
+from ._primary import Primary
 
-class Schema(_Datalike):
-    ...
+from .scalar import *
 
 
 ###############################################################################
