+++ everest/ptolemaic/datalike/primary/_primary.py
@@ -3,14 +3,11 @@
 ###############################################################################
 
 
-from .. import _classtools, _everestutilities
+from . import _Datalike
 
-from .. import Datalike as _Datalike
-from .. import functional as _functional
 
-from ._schema import Schema
-
-from .scalar import *
+class Primary(_Datalike):
+    ...
 
 
 ###############################################################################
