+++ everest/ptolemaic/datalike/primary/scalar/__init__.py
@@ -3,11 +3,16 @@
 ###############################################################################
 
 
-from ..exceptions import * # pylint: disable=W0401,W0614
+from .. import _classtools, _everestutilities
 
+from .. import _functional
+from .. import Primary as _Primary
 
-class FunctionalException(DatalikeException):
-    ...
+from ._scalar import Scalar
+
+from .integer import *
+from .real import *
+from .text import *
 
 
 ###############################################################################
