+++ everest/ptolemaic/datalike/primary/scalar/exceptions.py
