+++ everest/ptolemaic/datalike/primary/scalar/text.py
