+++ everest/ptolemaic/datalike/secondary/armatures/brace.py
@@ -3,16 +3,11 @@
 ###############################################################################
 
 
-from .. import _classtools, _everestutilities
+from . import Armature as _Armature
 
-from .. import _functional
-from .. import Schema as _Schema
 
-from ._scalar import Scalar
-
-from .integer import *
-from .real import *
-from .text import *
+class Brace(_Armature):
+    ...
 
 
 ###############################################################################
