+++ everest/ptolemaic/datalike/secondary/functional/applicator.py
