+++ everest/ptolemaic/datalike/secondary/functional/operation.py
